feat: change thrown error be named ValidationError
When catching errors thrown by validators the name of the error is not
modified, this means the error can be confusing. For example superstruct
throws errors named StructError, so the final error is logged as
"StructError: Validation error at {model}.{action}: {msg}". The new
error format would result in "ValidationError: {model}.{action}: {msg}"